Create README.md
national cod ir 2142009085
name: muhammad mahdi saif litkouhi
sh:146
1354/02/25
father name: mohammad bagher
help me in
Dr. Hassan Jahid from the Ministry of Health was a former consultant to the Ministry of Health. He came here and uses this system for Shaina carpet houses. The women of the people all become invisible and go. Their husbands do not know and it has created problems for us for 15 years and 20 years. I have been tortured on the system.  I don't know how to control this, I am being harassed day and night, but I don't know what to doThe son of the same Mr. Hassan Jahid and the brother of Behrouz Atai were arrested for the crime of having a crime in the university.  They are going to take the child to their son's house for a vacation. Now I open this system from their mouth, they will tie it to me, then I will do whatever I can.  As for making a record, Mujahideen should use it. Let's do something so that I can show myself and introduce a system.My brother Mohammad Ali Seyfarid Kui went outside and Anna spoke and said that my brother should not be tortured by the IRGC, but he coordinated with my uncle without my knowledge, then he introduced me as the one who is being tortured the most, then I want them to destroy me.  And they should use this system. It was in the name of Imam Javad's brother. They should use the benefits and no matter what I do, I can't get rid of it. Helena, let's follow up on this letter. I'll send you a follow-up number. Find the serial number of my device. It has three Wi-Fi.  They close the monkey, they use it for the environment and for biology. The Wi-Fi system is inside my body. They are abusing me. My GPS is everything. My brother left and was working for himself. Please introduce me. Tell me.  Follow up the dead Martyr Shahid MojahedHello, Mujahideen, you can hear my voice. You are abroad. I had a request from you, even though I was not a member of the Mujahideen, but you accidentally blocked my tobacco device and because I was born on the date of Imam Zaman's birthday, it was not my fault. It was not your fault.  I was born on this date, the date of birth of Imam Zaman, they did this to me, the torture machine was closed, they did not torture me the most, you know that I am not guilty, and blaming the army or the army was the Mujahideen themselves, I have a request from you, I will give all the torture  He accepted and I suffered most of this pain.  Let's test with their bodiesOf all the torture programs and he says that he shows the machine, it should be tested with their bodies to see if their bodies accept such torture, will it show the waves in their body cells in their blood or not?  does not deceiveHello, I wanted to announce that these gentlemen, whose names I mention, are involved in the murder of me and my father because of the inheritance that Mossadegh Mossadegh wanted from the Saudi oil company, and because Iran has chosen me as the heir, I want the money that Mr. Mossadegh is asking for.  Take them from Saudi Arabia. After these names, friends of Ghulam Mohammad Hashmati, Fariborz Bagheri and Saeed Abdullah Khatami are among the other people who work with this family and relatives.  follow up the person who was the main one and it was me, I want to show myself, they are blocking him, they are making me show myself, I want him to take action after my death withoutMr. Hassan Kabuli, Hossein Kabuli, Mahmoud Dadashi, Mitsubishi agricultural equipment that they have in Tehran, their blood is in Tehran, Yousef Abad district, after that, they are all involved in this case to get money from Mr. Mohammad Mahdi Saifi, change to Mohammad Saif, my name is Mohammad Mahdi, change to Mohammad Seyed Mohammad Saifi.  You don't have any medicine, I took all my inheritance from the beginning to the end, and I used to do this as a child, the rich children did it and extorted themselves, and they used to extort money from me.  He cooperates. There are a number of policemen who cooperate with them. They don't care about me. I am doing it because I want to get out of Iran. I want to kill them.